use F= 0.1, constant selectivity for fishery
Also modify the lbin vector. Note that F was change to 0.1 from 0.2, b/c of conversion issues suggesting that 0.2 (and even 0.15) was too high and caused the population to crash. This means F is different than piner et al...I am not sure exactly why, maybe different q was used?